MavenResolver parse coordinates properly
Added a small Artifact class that takes maven coords and provides methods to get the various parts. This makes it easier for MavenResolver to extract those.

Followed the gradle-style maven coordinates which are slightly different from the aether coordinates used internally by maven.

This is a precursor to some upcoming MultiMC memes...